+++ include/zarr.types.h
@@ -66,7 +66,6 @@ extern "C"
         ZarrCompressor_None = 0,
         ZarrCompressor_Blosc1,
         ZarrCompressor_Zstd,
-        ZarrCompressor_Lz4,
         ZarrCompressorCount
     } ZarrCompressor;
 
@@ -76,7 +75,6 @@ extern "C"
         ZarrCompressionCodec_BloscLZ4,
         ZarrCompressionCodec_BloscZstd,
         ZarrCompressionCodec_Zstd,
-        ZarrCompressionCodec_Lz4,
         ZarrCompressionCodecCount
     } ZarrCompressionCodec;