+++ src/streaming/compression.params.hh
@@ -2,11 +2,10 @@
 
 #include "blosc.compression.params.hh"
 #include "zstd.compression.params.hh"
-#include "lz4.compression.params.hh"
 
 #include <variant>
 
 namespace zarr {
 using CompressionParams =
-  std::variant<BloscCompressionParams, ZstdCompressionParams, Lz4CompressionParams>;
+  std::variant<BloscCompressionParams, ZstdCompressionParams>;
 } // namespace zarr